Suggest CommandResult class
1. Instead of using System,out.println() to print every sentence
or message, add every message to a String "feedbackToUser" in each command

2. Every Command returns a new CommandResult(feedbackToUser)

3. Parser will then print the feedbackToUser string

4. When testing, instead of using ByteArrayOutputString, simply
return commandResult.feedbackToUser for the output message.

5. You can use the TestUtil getOutputMessage method to execute command
and get their output message

6. You can use formatting methods in Messages class to format lines
with line separators, line dividers. For example if the last line can
be formatted with "formatLastLineDivider". If you have any other
formats that seem to be repeating in your code add a method to the
Messages class to use